refactor: write memory storage
