+++ pkg/domain/accounts/check_accounts.go
@@ -1,0 +1,5 @@
+package accounts
+
+func (au AccountUseCase) CheckAccounts(id ...string) error {
+	return au.storage.CheckAccountsByID(id)
+}
