+++ pkg/domain/accounts/get_account_by_id.go
@@ -1,0 +1,9 @@
+package accounts
+
+import (
+	"exemplo.com/pkg/domain/entities"
+)
+
+func (au AccountUseCase) GetAccountByID(id string) (entities.Account, error) {
+	return au.storage.FindByID(id)
+}
