+++ pkg/domain/accounts/get_accounts.go
@@ -1,0 +1,9 @@
+package accounts
+
+import (
+	"exemplo.com/pkg/domain/entities"
+)
+
+func (au AccountUseCase) GetAccounts() []entities.Account {
+	return au.storage.JoinAccounts()
+}
