+++ pkg/domain/accounts/get_balance_by_id.go
@@ -1,0 +1,6 @@
+package accounts
+
+func (au AccountUseCase) GetBalanceByID(id string) (int, error) {
+	return au.storage.FindBalanceByID(id)
+
+}
