+++ pkg/domain/accounts/new_account.go
@@ -1,0 +1,15 @@
+package accounts
+
+import (
+	"exemplo.com/pkg/store"
+)
+
+type AccountUseCase struct {
+	storage store.AccountStorage
+}
+
+func NewAccountUseCase(storage store.AccountStorage) AccountUseCase {
+	return AccountUseCase{
+		storage: storage,
+	}
+}
