+++ pkg/domain/accounts/update_account_balance.go
@@ -1,0 +1,24 @@
+package accounts
+
+import (
+	"errors"
+)
+
+var (
+	ErrBalanceLessOrEqualZero = errors.New("balance account less or equal zero")
+)
+
+func (au AccountUseCase) UpdateAccountBalance(id string, balance int) error {
+	account, err := au.GetAccountByID(id)
+	if err != nil {
+		return err
+	}
+	if balance < 0 {
+		return ErrBalanceLessOrEqualZero
+	}
+
+	account.Balance = balance
+	au.storage.UpdateStorage(account.ID, account)
+
+	return nil
+}
