+++ pkg/domain/transfers/make_transfer.go
@@ -10,16 +10,6 @@ var (
 	ErrToCallNewTransfer = errors.New("error to call function NewTransfer")
 )
 
-type TransferUseCase struct {
-	storage map[string]entities.Transfer
-}
-
-func NewTransferUseCase(storage map[string]entities.Transfer) TransferUseCase {
-	return TransferUseCase{
-		storage: storage,
-	}
-}
-
 func (tu *TransferUseCase) MakeTransfer(originID, destinationID int, amount int) (entities.Transfer, error) {
 
 	transfer, err := entities.NewTransfer(originID, destinationID, amount)
@@ -28,7 +18,7 @@ func (tu *TransferUseCase) MakeTransfer(originID, destinationID int, amount int)
 		return entities.Transfer{}, ErrToCallNewTransfer
 	}
 
-	tu.storage[transfer.ID] = transfer
+	tu.storage.UpdateTransferStorage(transfer.ID, transfer)
 
 	return transfer, nil
 }
