+++ pkg/domain/transfers/make_transfer_test.go
@@ -4,13 +4,14 @@ import (
 	"testing"
 
 	"exemplo.com/pkg/domain/entities"
+	"exemplo.com/pkg/store"
 )
 
 func TestAccountUseCase_MakeTransfer(t *testing.T) {
 
 	t.Run("Should return a transfer", func(t *testing.T) {
 
-		storage := make(map[string]entities.Transfer)
+		storage := store.NewTransferStorage()
 		TransferUsecase := NewTransferUseCase(storage)
 
 		MakeTransfer, err := TransferUsecase.MakeTransfer(1, 2, 10)
