+++ pkg/domain/transfers/new_transfer_use_case.go
@@ -1,0 +1,15 @@
+package transfers
+
+import (
+	"exemplo.com/pkg/store"
+)
+
+type TransferUseCase struct {
+	storage store.TransferStorage
+}
+
+func NewTransferUseCase(storage store.TransferStorage) TransferUseCase {
+	return TransferUseCase{
+		storage: storage,
+	}
+}
