+++ pkg/store/get_cpf.go
@@ -1,0 +1,10 @@
+package store
+
+func (s AccountStorage) GetCPF(cpf string) error {
+	for _, storedAccount := range s.storage {
+		if storedAccount.CPF == cpf {
+			return ErrExistingCPF
+		}
+	}
+	return nil
+}
