+++ pkg/store/join_accounts.go
@@ -1,0 +1,11 @@
+package store
+
+import "exemplo.com/pkg/domain/entities"
+
+func (s AccountStorage) JoinAccounts() []entities.Account {
+	var AccountsList []entities.Account
+	for _, value := range s.storage {
+		AccountsList = append(AccountsList, value)
+	}
+	return AccountsList
+}
