+++ pkg/store/new_transfer_storage.go
@@ -1,0 +1,14 @@
+package store
+
+import "exemplo.com/pkg/domain/entities"
+
+type TransferStorage struct {
+	storage map[string]entities.Transfer
+}
+
+func NewTransferStorage() TransferStorage {
+	sto := make(map[string]entities.Transfer)
+	return TransferStorage{
+		storage: sto,
+	}
+}
