+++ pkg/store/update_storage.go
@@ -1,0 +1,7 @@
+package store
+
+import "exemplo.com/pkg/domain/entities"
+
+func (s AccountStorage) UpdateStorage(id string, account entities.Account) {
+	s.storage[id] = account
+}
