+++ pkg/store/update_transfer_storage.go
@@ -1,0 +1,7 @@
+package store
+
+import "exemplo.com/pkg/domain/entities"
+
+func (t TransferStorage) UpdateTransferStorage(id string, transfer entities.Transfer) {
+	t.storage[id] = transfer
+}
